+++ scripts/variables.sh
@@ -1,0 +1,23 @@
+# stored search variable prefix
+COPYCAT_VAR_PREFIX="@copycat_search"
+
+# basic search
+default_copycat_search_key="/"
+copycat_search_option="@copycat_search"
+
+# git special search
+default_git_search_key="C-g"
+copycat_git_search_option="@copycat_git_special"
+
+# regular searches
+default_file_search_key="C-f"
+copycat_file_search_option="@copycat_file_search"
+
+default_url_search_key="C-u"
+copycat_url_search_option="@copycat_url_search"
+
+default_digit_search_key="C-d"
+copycat_digit_search_option="@copycat_digit_search"
+
+default_ip_search_key="M-i"
+copycat_ip_search_option="@copycat_ip_search"